pyproject.toml
[project]
name = "adaptive-cyber-crisis-environment"
version = "0.1.0"
description = "Deterministic OpenEnv-compatible cyber crisis simulation environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core",
  "fastapi",
  "uvicorn",
  "pydantic",
  "openai",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
